Changed the i_x variables back to doubles

+++ src/main/main.c
@@ -4,22 +4,22 @@
 int main(int argc, char *argv[]){
 	mpf_set_default_prec(5);
 
-	float i_a, i_b, i_c; // Will be parsed from input
+	double i_a, i_b, i_c; // Will be parsed from input
 	mpf_t a, b, c; // Will be worked with throughout
 
 	// mpf_set_d(num, -4);
 
 	if(argc<4){
 	    printf("Please enter 3 floating-point numbers with a newline after each:\n");
-        scanf("%f",&i_a);
-        scanf("%f",&i_b);
-        scanf("%f",&i_c);
+        scanf("%lf",&i_a);
+        scanf("%lf",&i_b);
+        scanf("%lf",&i_c);
 	} 
 
 	else{
-	    sscanf(argv[1], "%f", &i_a);
-	    sscanf(argv[2], "%f", &i_b);
-	    sscanf(argv[3], "%f", &i_c);
+	    sscanf(argv[1], "%lf", &i_a);
+	    sscanf(argv[2], "%lf", &i_b);
+	    sscanf(argv[3], "%lf", &i_c);
 	}
 
 	mpf_init(a);
